abstracted fastapi and socketio functionality to services, removed old commented code
